Change all cases then used int type to int64

+++ sources/fourChannel/ResponsePosts.go
@@ -4,6 +4,6 @@ type ResponsePosts struct {
 	Posts []struct {
 		Filename      string `json:"filename"`
 		FileExtension string `json:"ext"`
-		FileId        int    `json:"tim"`
+		FileId        int64    `json:"tim"`
 	}
 }
