+++ sources/types/File.go
@@ -4,5 +4,5 @@ type File struct {
 	Name     string `json:"name"`
 	Path     string `json:"path"`
 	Preview  string `json:"preview"`
-	ThreadId int    `json:"threadId"`
+	ThreadId int64    `json:"threadId"`
 }
